backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic[email]
passlib[bcrypt]
bcrypt==4.0.1
python-jose
pymongo
certifi
google-genai
markdown2
motor
PyPDF2
python-docx
